Made it possible to open the user guide in deployed versions.
The code should work in deployed versions on PC and Mac, but not for linux.
The code would need to be extended in order to create a deployed version
for linux, but at the moment there are no plans to create such a version.

When new deployed versions have been created, issue #2 will be solved by
this commit.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>