Added breakling of all clusters when bipartite matching is not performed.
Previously, all of the cells in a cluster had the outline of the un-broken
cluster when no bipartite matching was preformed. This is usually not
wanted and it causes problems in the TRIC dataset in the 2019 Cell Tracking
Challenge, as the CTC format requires that each pixel is assigned to at
most one cell.

Signed-off-by: Klas Magnusson <klas.magnusson@raysearchlabs.com>