Added the setting TrackRelinkSelectedCells.
The setting makes it possible to select whether or not tracks with a CTC
ground truth should be extended to the end of the image sequence by
linking them to track fragments in subsequent frames. The relinking was
hard-coded for the DRO dataset in CTC 2015 and removed in CTC 2019 when the
TRIC dataset was added. The performance on the DRO dataset did however go
down, so now relinking is done for DRO but not for TRIC. The setting only
makes a difference for the DRO and TRIC datasets, where there is a marker
in the first frame for each cell that should be tracked.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>