Made BreakClusters handle time gaps and parents with a single child.
Parents with a single child occur in the CTC data where they are used to
represent time gaps when a cell disappears and reappears later. Handling of
such cells was added.

I also changed the existing code so that it can handle time gaps between
the last frame of the parent cell and the first frame of the child cell.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>