Added display of cell indices to the manual correction GUI.
An index property has been added to the Cell class. The property should
always be equal to the index of the cell in the array of cells that it has
been saved (or will be saved) to. The index is updated by the SortCells
method. For old tracking data that was created before this code change, the
indices are generated in the LoadCells function.

The dispay of indices in the manual correction GUI (and fiber correction
GUI) is toggled by pressing Alt+N. By default, the indices are not shown.
The text color is the same as the color of false positive objects.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>