Optimized memory usage for segmentation of the TRIF dataset.
Convert only the block to be segmented into a double-array instead of
converting the whole volume and then picking a sub-volume.

Clear variables after their last use to free up memory.

Stop caching smoothed integer images when segmenting blocks. The blocks
have different sizes, so the cached images are replaced a lot. The run
time for the smoothing is relatively small.

Avoid storing outputs which are not needed.