Optimized the segmentation parameters for the MSC datasets.
The the optimization performed both segmentation and tracking. This was
done as the final result depends a lot on how watershed fragments of
background objects are incorporated into the cells. The mean of SEG and TRA
was used as utility function. The results did not look very promising on
the challenge dataset, so this commit will be reverted.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>