A new migration score based on distances between blob fragments.
Instead of using distances between centroids of blob pairs, this function
uses distances between the centroid of the smaller blob the centorid of
a fragment of the larger blob. The fragemnt is composed of the N pixels
which are closest to the centorid of the smaller blob, where N is the
area of the smaller blob.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>