Added parallel processing to the CTC program.
The parallelization is performed in the segmentation step. The number of
cores used is the total amount of RAM divided by the RAM required for a
single frame. The amount of RAM required for a single frame is set
conservatively so that the memory will be sufficient on a computer with 32
or more GB of RAM.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>