Used the available RAM instead of the total RAM in the CTC program.
This is safer as other programs may be running on the computer. It is also
more likely to work on computers with less than 32 GB of RAM as a computer
with very limited RAM will not have a lot of RAM left once the OS and
Matlab have taken what they need. I also made sure that at least one core
is used.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>