Adjusted C. elegans settings for a shorter second challenge sequence.
In CTC V, the second challenge sequence in the C. elegans data has been
shortened from 190 to 140 frames. Some of the settings in the C. elegans
dataset vary linearly with the frame index. To get the same settings in the
frames of the shorter sequence, the settings of the last frame had to be
changed. The new settings were computed by using linear interpolation to
get the settings in frame 140 of the old data.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>