Added export of segmentation ground truths in 3D.
Previously, an error occurred if the user tried to export segmentation
ground truths for 3D datasets. Now it is possible to export segmentation
ground truths in the same format as in the Fluo-N3DH-SIM+ dataset. It is
still not possible to export segmentation ground truths in the format used
in the other challenge datasets.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>