Fixed bug in loading of CTC results in 3D datasets with unstacked z-stacks.
The ImageData object used to read CTC tracking results took the zStacked
property from the image data, but it should always be true as the CTC
format always uses stacked tifs for z-stacks. The CTC format is used as a
fallback if the tracking results cannot be saved as mat-files.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>