Removed type conversions from mwSize* to int* in mex files.
The 64-bit mex API uses the type mwSize (const long unsigned int) for some
inputs and outputs. The BaxterAlgorithms used int instead. It seems like it
worked with the C++ compiler I used, but it does not seem to work with GCC
when people are trying to compile the mex files for Linux systems. It could
also be that I used an older version of the Matlab mex API when I compiled.
Hopefully, these changes will make it possible to compile using all
compilers, including Linux compilers. I have not yet checked in the new
compiled mex-files, as I want to make sure that the solution is correct
first.

I also added "#include <algorithm>" to Variable.cpp to make the code
compile using the Microsoft Visual C++ 2017 compiler.

Signed-off-by: Klas Magnusson <klas.magnusson@raysearchlabs.com>