Switched back to unoptimized segmentation settings for the TRIF datasets.
The settings are taken from 186c7680803c187bc71705aacf227a0cc3f81c28. Most
of the settings are the same as the ones used for TRIC, but some have been
changed to make the calculations more efficient.

In the challenge of 2020, I got much worse results for the TRIF datasets
than on the TRIC datasets. That should not be the case and is probably due
to overfitting to the training data in the segmentation optimization. This
year, I am planning to use the unoptimized segmentation settings instead.