Updated the starting points for the segmentation optimization.
I updated the clean settings that are used as starting points for the
segmentation optimization. I changed the settings for the bandpass
segmentation so that it gives smoother contours. This seems to work a
little better for most image sequences. I also added a watershed transform
based on the bandpass filtered image and a second watershed based on the
region shape. Both watersheds use HMax transforms to avoid
over-segmentation. I changed SegWSmooth from 7 to 0, as that parameter is
used on few CTC datasets. I also changed the minimum area from 40 to 0 to
avoid removing small cells. I added the background subtraction for MuSC and
HSC as this will make the performance evaluation more meaningful. Now, the
sequence names do not include the experiment names, for consistency with
the other settings files.

Signed-off-by: Klas Magnusson <klasmagnus@gmail.com>