Made it possible to have multiple settings files and/or link files.
Now, a settings link file can be given as input to ReadSettings and
WriteSettings. Previously, it was only possible to give a settings file as
input and the settings link file was only used when an experiment folder
was given as input.

I have also added an assert which verifies that settings files or link
files exist when ReadSettings is called. I also added a warning if a
settings file does not contain settings for the desired image sequence. I
did not use an assert in this case because that would make it hard to
add additional image sequences to an existing experiment.