Removed transmission microscopy data from joint trainng script.
The results produced for the transmission microscopy datasets will not be
usable in biological applications. It is probably better to show good
results on fluorescence data than to optimize non-usable results to get
a better average score and sacrifice performance on fluorescene data for
that. The optimization will also run faster with fewer datasets.

Signed-off-by: Klas Magnusson <klas.magnusson@raysearchlabs.com>