Switched to optimization of the relaxed SEG measure.
This was done primarily because it was not possible to optmize the SEG
measrue on the A549 dataset. This was becasue the optimization increased
the BPSegBgFactor until there were no cells left. The problem was that
SegWHMax was too small now that the datasets are rescaled to have a range
between 0 and 1. The optimization did not have time to increase SegWHMax
enough before all cells disappeared due to the increase in BPSegBgFactor.
In this optimization, the SEG measure is 0 all the time but the relaxed SEG
measure is not and therefore prevents BPSegBgFactor from increasing. The
relaxed SEG measure should also be more suitable for optimization in other
situations too, as it is smoother and is non-zeros for almost all
segmentations.