Major Improvements based on DataScientists Conversation
Made some major formatting changes! the places marked with "##" (just ctrl+F for ##) are notes on things that make good specific things to be working on, particularlyy for michale
Also added feature to automatically save the exacdt file that was run everytime we run teh code. This should help with record keeping and everythnig a ton.