Almost Working GUI
All code from the new script is put into the GUI. The only thing left to do to get the GUI to run is to set up the ImageAnalyses cell array variable correctly. Also still need to change the layout of how the image planes are selected for each function.